fixed some ui  bullshit

+++ Backend/src/main/java/org/anuj/miniprojectfintech/DashBoard/StudentSummary.java
@@ -3,6 +3,8 @@ package org.anuj.miniprojectfintech.DashBoard;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 
+import java.math.BigDecimal;
+
 @Getter
 @AllArgsConstructor
 public class StudentSummary {
@@ -12,6 +14,6 @@ public class StudentSummary {
     private String year;
     private String branch;
     private Boolean active;
-    private long pendingFees;
-    private long totalFees;
+    private BigDecimal pendingFees;
+    private BigDecimal totalFees;
 }
